fix compilation of sdl-deltalake unit tests
