+++ sdl-core/src/test/scala/io/smartdatalake/testutils/custom/TestCustomDfManyTypes.scala
@@ -1,7 +1,7 @@
 /*
  * Smart Data Lake - Build your data lake the smart way.
  *
- * Copyright © 2020 Schweizerische Bundesbahnen SBB (<https://www.sbb.ch>)
+ * Copyright © 2019-2021 ELCA Informatique SA (<https://www.elca.ch>)
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -16,26 +16,17 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package io.smartdatalake.config
+package io.smartdatalake.testutils.custom
 
+import io.smartdatalake.testutils.TestUtil.dfManyTypes
 import io.smartdatalake.workflow.action.customlogic.CustomDfCreator
-import org.apache.spark.sql.types.{IntegerType, StructField, StructType}
+import org.apache.spark.sql.types.StructType
 import org.apache.spark.sql.{DataFrame, SparkSession}
 
-class TestCustomDfCreatorWithSchema extends CustomDfCreator {
+class TestCustomDfManyTypes extends CustomDfCreator {
+  override def exec(session: SparkSession, config: Map[String, String]): DataFrame = dfManyTypes
 
-  override def exec(session: SparkSession, config: Map[String, String]): DataFrame = {
-      import session.implicits._
-      val rows: Seq[(Some[Int], String)] = Seq((Some(0),"Foo!"),(Some(1),"Bar!"))
-      val myDf: DataFrame = rows.toDF("num","text")
-      myDf.show(false)
-      myDf
-  }
-
-  override def schema(session: SparkSession, config: Map[String, String]): Option[StructType] = {
-    // this schema does not match the schema of the DataFrame returned by exec
-    Option(new StructType(Array(new StructField("num", IntegerType))))
-  }
+  override def schema(session: SparkSession, config: Map[String, String]): Option[StructType] = Option(dfManyTypes.schema)
 
   override def equals(obj: Any): Boolean = getClass.equals(obj.getClass)
 }
