+++ sdl-core/src/test/scala/io/smartdatalake/testutils/custom/TestCustomDfNonUniqueWithNullCreator.scala
@@ -1,7 +1,7 @@
 /*
  * Smart Data Lake - Build your data lake the smart way.
  *
- * Copyright © 2019-2020 ELCA Informatique SA (<https://www.elca.ch>)
+ * Copyright © 2019-2021 ELCA Informatique SA (<https://www.elca.ch>)
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -16,26 +16,17 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package io.smartdatalake.config
+package io.smartdatalake.testutils.custom
 
+import io.smartdatalake.testutils.TestUtil.dfNonUniqueWithNull
 import io.smartdatalake.workflow.action.customlogic.CustomDfCreator
+import org.apache.spark.sql.types.StructType
 import org.apache.spark.sql.{DataFrame, SparkSession}
 
-class TestCustomDfCreator extends CustomDfCreator {
-  /**
-   * function to create a dataFrame
-   *
-   * @param session : the Spark Session
-   * @param config  Input Config of action
-   * @return Map outputID -> dataFrame
-   */
-  override def exec(session: SparkSession, config: Map[String, String]): DataFrame = {
-      import session.implicits._
-      val rows: Seq[(Some[Int], String)] = Seq((Some(0),"Foo!"),(Some(1),"Bar!"))
-      val myDf: DataFrame = rows.toDF("num","text")
-      myDf.show(false)
-      myDf
-  }
+class TestCustomDfNonUniqueWithNullCreator extends CustomDfCreator {
+  override def exec(session: SparkSession, config: Map[String, String]): DataFrame = dfNonUniqueWithNull
+
+  override def schema(session: SparkSession, config: Map[String, String]): Option[StructType] = Option(dfNonUniqueWithNull.schema)
 
   override def equals(obj: Any): Boolean = getClass.equals(obj.getClass)
 }
