+++ sdl-core/src/test/scala/io/smartdatalake/workflow/dataobject/CustomFileDataObjectTest.scala
@@ -19,8 +19,8 @@
 
 package io.smartdatalake.workflow.dataobject
 
-import io.smartdatalake.config.TestCustomFileCreator
 import io.smartdatalake.testutils.DataObjectTestSuite
+import io.smartdatalake.testutils.custom.TestCustomFileCreator
 import io.smartdatalake.workflow.action.customlogic.CustomFileCreatorConfig
 import org.scalatest.Matchers
